+++ csrc/utils/math.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 // #include <torch/python.h>
+#include <cstdint>
 
 #include "exception.hpp"
 
@@ -16,8 +17,8 @@ static constexpr T align(const T& a, const T& b) {
     return ceil_div(a, b) * b;
 }
 
-static int get_tma_aligned_size(const int& x, const int& element_size) {
-    constexpr int kNumTMAAlignmentBytes = 16;
+static int64_t get_tma_aligned_size(const int64_t& x, const int64_t& element_size) {
+    constexpr int64_t kNumTMAAlignmentBytes = 16;
     DG_HOST_ASSERT(kNumTMAAlignmentBytes % element_size == 0);
     return align(x, kNumTMAAlignmentBytes / element_size);
 }
